+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "L-triominoes".
 - Tests for "Paper labyrinth".
 - Tests for "Minesweeper".
+- Tests for "The Grand Festival - I".
 
 ## [1.11.0] - 2022-07-31
 ### Added